fix import of user

+++ src/main/java/dev/lpa/goutbackend/wallet/models/UserWallet.java
@@ -1,18 +1,21 @@
 package dev.lpa.goutbackend.wallet.models;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
-import org.springframework.boot.autoconfigure.security.SecurityProperties.User;
+
 import org.springframework.data.annotation.Id;
 import org.springframework.data.jdbc.core.mapping.AggregateReference;
 import org.springframework.data.relational.core.mapping.Table;
+
+import dev.lpa.goutbackend.user.models.User;
 
 @Table("user_wallet")
 public record UserWallet(
     @Id Integer id,
     AggregateReference<User, Integer> userId,
     Instant lastUpdated,
-    Integer balance
+    BigDecimal balance
 ) {
 
 }
